m1 started

+++ Libasm/m0/ex2/main.c
@@ -1,11 +1,12 @@
 #include <stdio.h>
+#include <unistd.h>
 
-extern int ft_write(int fd, void *buf, int count);
+extern ssize_t ft_write(int fd, void *buf, size_t count);
 
 int main(void)
 {	
 	char *str = "Hello, World!\n";
-	int i = ft_write(1, str, 14);
-	printf("%d\n", i);
+	ssize_t i = ft_write(1, str, 14);
+	printf("%ld\n", i);
 	return 0;
 }
